+++ pages/index.js
@@ -121,7 +121,7 @@ function Home() {
   const imgSrcA = `https://source.unsplash.com/1024x768?japan,city,wallpaper`;
   const imgSrcB = `https://source.unsplash.com/collection/4655931`;
   const imgSrcC = `https://source.unsplash.com/collection/8593901`;
-  const specific = `https://source.unsplash.com/alY6_OpdwRQ`;
+  const specific = `https://source.unsplash.com/layMbSJ3YOE`;
 
   return (
     <Container>