+++ components/tsucounters/tsuCountersHelper.js
@@ -1,10 +1,4 @@
-function createRandomInteger(min, max) {
-  let int = Math.round(Math.random() * (max - min));
-  if (int === 0) {
-    return 1;
-  }
-  return int;
-}
+import { createRandomInteger } from '../../utils/random';
 
 function randomItem() {
   const counterTypes = [